+++ CHANGELOG.md
@@ -1,5 +1,10 @@
 # Changelog
 
+### 0.0.4
+- added showing version
+- fixed default path to custom templates folder
+- added log for successfully set options
+
 ### 0.0.3
 - added support to set path for custom templates
 - fixed to generate custom templates